backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.34
pydantic[email]==2.9.2
pydantic-settings==2.5.2
email-validator==2.2.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-multipart==0.0.9
httpx==0.27.2
websockets==12.0
alpaca-py==0.30.1
pytz==2024.2
apscheduler==3.10.4
twscrape==0.13.1

